Milestone2-V2
The if-else statements and while loops of the hangman game. Both if statements are in the while loop. The first if statement checks if the input from the user/player is a single character and if it is an alphabet. If both conditions are not met, then the game tells the user/player the input is invalid and a new input must be entered.  The second if statement checks if the input - called guess - is in the work from the list. If not, the game tells the user/player the input is incorrect. The while loop repeated if the input does not work for both if statements.